requirements.txt
selenium
chromedriver_autoinstaller
opencv-python
numpy
easyocr
gspread
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
